# api/cassetteos/openapi.yaml
openapi: 3.0.3
info:
  title: CasaOS API
  version: v2
  description: |
    <picture>
        <source media="(prefers-color-scheme: dark)" srcset="https://raw.githubusercontent.com/IceWhaleTech/logo/main/casaos/casaos_banner_dark_night_800px.png">
        <source media="(prefers-color-scheme: light)" srcset="https://raw.githubusercontent.com/IceWhaleTech/logo/main/casaos/casaos_banner_twilight_blue_800px.png">
        <img alt="CasaOS" src="https://raw.githubusercontent.com/IceWhaleTech/logo/main/casaos/casaos_banner_twilight_blue_800px.png">
    </picture>

    CasaOS API provides miscellaneous methods for different scenarios.

    For issues and discussions, please visit the [GitHub repository](https://github.com/mayumigit/CasaOS) or join [our Discord](https://discord.gg/knqAbbBbeX).
servers:
  - url: /v2/casaos
...
security:
  - access_token: []
paths:
...
  /wifi:
    get:
      summary: Get current WiFi status
      operationId: getWiFiStatus
      tags:
        - WiFi
      responses:
        '200':
          description: WiFi status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WiFiStatus'
    post:
      summary: Set WiFi configuration
      operationId: setWiFiConfig
      tags:
        - WiFi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WiFiConfig'
      responses:
        '200':
          description: WiFi configuration applied
  /wifi/ap-mode:
      post:
        summary: Switch to AP mode
        operationId: setWiFiAPMode
        tags:
          - WiFi
        responses:
          '200':
            description: Switched to AP mode successfully
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: string
                      example: ok
          '500':
            description: Failed to switch to AP mode
  /wifi/ap-ssid:
      get:
        summary: Get AP SSID
        operationId: getWiFiAPSSID
        tags:
          - WiFi
        responses:
          '200':
            description: Get AP SSID successfully
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: string
                      example: ok
          '500':
            description: Failed to get AP SSID
...
components:
  securitySchemes:
    access_token:
      type: apiKey
      in: header
      name: Authorization
...
  schemas:
...
    WiFiStatus:
      type: object
      properties:
        mode:
          type: string
          description: WiFi mode (e.g., client or ap)
        ssid:
          type: string
          description: Currently connected SSID
        ip_address:
          type: string
          description: Assigned IP address
    WiFiConfig:
      type: object
      required:
        - mode
      properties:
        mode:
          type: string
          description: WiFi mode to switch to
          enum: [wifi-client, ap]
        ssid:
          type: string
          description: SSID to connect to (only in client mode)
        password:
          type: string
          description: WiFi password (only in client mode)